fix: user Type 오류 해결 2
